Move navbar component
